fix: playbacklistener mount
reverts it back to PlaybackService.ts.
what we wanted to achieve (updated metadata) can be done with a vanilla store so lets do that instead. usePlayback can house the actual update as it needs access to useNoxSetting, its fine as where it is
